Plugin Curvature Skeleton - fixes and changes in descriptions
The mesh is now correctly checked before converting it to a CGAL mesh, not requiring the post-conversion test anymore. The QualitySpeed and MediallyCentering Tradeoff values do now follow the same values used in the academic paper. Some of the descriptions of the tool are now changed, including more details and a reference to the original paper.