Fixed Branch Extender, now correctly extending all branches. Minor refactorings.
Vertices that seem to be collapsing on the tip of a branch might not be linked to the tip of the branch on the skeleton, but can be found on the parent vertices of the tip too! The branch extender now looks at those vertices too, extending even not previously extended branches.