Changed collapse edges for root node
Edges incident on the root node are collapsed only if the collapsing happens on the root node rather than not collapsing the edge if it contains the root node even as collapse position. The algorithm should not be different because edges were already collapsed towards the root node, but now it makes more sense.