Changes to Generation of Logical Tree from Skeleton
The tree branches are now collapsed with a different criteria:
From the smallest to the longest, the edges are deleted then, if the collapsing vertex is one manifold, it is again collapsed.
The edges are tested on their new length after their base vertex is collapsed instead of their old length.
The last passage to remove the one manifold vertices is removed as it is not necessary anymore.